fear: add portfolio user service

+++ portfolio-service/src/main/java/com/chan/investment/portfolioservice/jpa/service/PortfolioUserService.java
@@ -4,31 +4,23 @@ import com.chan.investment.portfolioservice.exception.EntityNotFoundException;
 import com.chan.investment.portfolioservice.jpa.dto.PortfolioDTO;
 import com.chan.investment.portfolioservice.jpa.entity.Portfolio;
 import com.chan.investment.portfolioservice.jpa.repository.PortfolioRepository;
+import jakarta.ws.rs.ForbiddenException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
-public class PortfolioService {
+public class PortfolioUserService {
 
     private final PortfolioRepository portfolioRepository;
 
-    public PortfolioService(PortfolioRepository portfolioRepository) {
+    public PortfolioUserService(PortfolioRepository portfolioRepository) {
         this.portfolioRepository = portfolioRepository;
     }
 
-    public PortfolioDTO getPortfolioById(Long id) {
-        Optional<Portfolio> portfolio = portfolioRepository.findById(id);
-        if (portfolio.isEmpty()) {
-            throw new EntityNotFoundException("Portfolio with id: " + id + " not found");
-        }
-        return PortfolioDTO.fromEntity(portfolio.get());
-    }
-
-    public List<PortfolioDTO> getPortfolios() {
-        List<Portfolio> portfolios = portfolioRepository.findAll();
+    public List<PortfolioDTO> getPortfolios(String username) {
+        List<Portfolio> portfolios = portfolioRepository.findAllByUsername(username);
         if (portfolios.isEmpty()) {
             throw new EntityNotFoundException("Portfolios not found");
         }
@@ -44,14 +36,21 @@ public class PortfolioService {
     }
 
     public PortfolioDTO updatePortfolio(Long id, PortfolioDTO portfolioDTO, String username) {
-        this.getPortfolioById(id);
+        Portfolio portfolio = portfolioRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Portfolio with id: " + id + " not found"));
+        if(!portfolio.getUsername().equals(username)){
+            throw new ForbiddenException("You are not allowed to update this portfolio");
+        }
         portfolioDTO.setId(id);
-        Portfolio portfolio = PortfolioDTO.toEntity(portfolioDTO, username);
-        Portfolio savedPortfolio = portfolioRepository.save(portfolio);
+        Portfolio updatePortfolio = PortfolioDTO.toEntity(portfolioDTO, username);
+        Portfolio savedPortfolio = portfolioRepository.save(updatePortfolio);
         return PortfolioDTO.fromEntity(savedPortfolio);
     }
 
-    public void deletePortfolio(Long id) {
+    public void deletePortfolio(Long id, String username){
+        Portfolio portfolio = portfolioRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Portfolio with id: " + id + " not found"));
+        if(!portfolio.getUsername().equals(username)){
+            throw new ForbiddenException("You are not allowed to update this portfolio");
+        }
         portfolioRepository.deleteById(id);
     }
